backend/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
gunicorn==20.1.0
numpy==1.24.1
pandas==1.5.2
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
